Add support for authentication
By defining a password in the environment variable `PASSWORD`
a cluster will be configured to require a client to authenticate itself.
The client needs to send the `AUTH` command followed by the password
before being able to access the cluster.

This is somewhat based on #94 but avoiding the compability problem
of having empty passwords in early 6.0 releases.